+++ changelog.md
@@ -14,6 +14,8 @@ Codion Change Log
 - Unit tests improved.
 ### is.codion.framework.db.core
 - Unit tests improved.
+### is.codion.framework.db.local
+- Unit tests improved.
 
 ## 0.18.39
 ### is.codion.common.core